dom/tests/mochitest/mouselock/test_navigatorPointer.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Test for Bug 92264</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
  <p id="display"></p>

  <div id="content" style="display: none">
  </div>

  <pre id="test">
    <script class="testbody" type="text/javascript">

      SimpleTest.waitForExplicitFinish();

      var pointer = navigator.pointer;

      ok(pointer, "expected navigator.pointer to exist.");

      ok(pointer instanceof MouseLockable, "expected navigator.pointer to be a MouseLockable.");

      ok(pointer.lock, "MouseLockable should have lock!.");
      ok(pointer.unlock, "MouseLockable should have unlock.");
      ok(pointer.islocked, "MouseLockable should have islocked.");

      is(pointer.islocked(), false, "Mouse Lock should initially be false.");

      SimpleTest.finish();

    </script>
  </pre>
</body>
</html>
